сделал refactoring приложения, также получил токен для дальнейшей загрузки данных из api
