+++ app/src/main/java/ru/paramonov/vknewsclient/presentation/screens/comments/CommentsScreenState.kt
@@ -1,4 +1,4 @@
-package ru.paramonov.vknewsclient.screens.comments
+package ru.paramonov.vknewsclient.presentation.screens.comments
 
 import ru.paramonov.vknewsclient.domain.FeedPost
 import ru.paramonov.vknewsclient.domain.PostComment
