+++ app/src/main/java/ru/paramonov/vknewsclient/presentation/screens/login/AuthViewState.kt
@@ -1,4 +1,4 @@
-package ru.paramonov.vknewsclient.screens.login
+package ru.paramonov.vknewsclient.presentation.screens.login
 
 sealed class AuthViewState {
