+++ app/src/main/java/ru/paramonov/vknewsclient/samples/ActivityResultApi.kt
@@ -1,4 +1,4 @@
-package ru.paramonov.vknewsclient
+package ru.paramonov.vknewsclient.samples
 
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -18,7 +18,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
